Use MemoryStoreProvider for `@AiService` (#236)
Close: #234

Previously, we were invoking builder.chatMemory which means that you use the same (shared) ChatMemory for all users/conversations.

AIServices::chatMemoryProvider javadoc:

If you prefer to use the same (shared) ChatMemory for all users/conversations, configure a chatMemory instead.
Either a ChatMemory or a ChatMemoryProvider can be configured, but not both simultaneously.
However, using chatMemory fails with a NPE for such an @AiService

 @AiService
    interface Assistant {
        String chat(@MemoryId String memoryId, @UserMessage String message);
 }
This Pull-Request changes to use chatMemoryProvider instead.